added the CollisionsUtil class
