+++ two/src/main/java/com/wizardlybump17/physics/two/shape/Circle.java
@@ -5,6 +5,7 @@ import com.wizardlybump17.physics.two.intersection.circle.CircleToCircleIntersec
 import com.wizardlybump17.physics.two.intersection.rectangle.RectangleToCircleIntersection;
 import com.wizardlybump17.physics.two.position.Vector2D;
 import com.wizardlybump17.physics.two.shape.rotating.RotatingPolygon;
+import com.wizardlybump17.physics.two.util.CollisionsUtil;
 import com.wizardlybump17.physics.util.MathUtil;
 import org.jetbrains.annotations.NotNull;
 
@@ -60,7 +61,7 @@ public class Circle extends Shape {
 
                 yield MathUtil.square(xDistance - width) + MathUtil.square(yDistance - height) <= MathUtil.square(radius);
             }
-            case RotatingPolygon polygon -> RotatingPolygon.collidesWith(polygon.getRotatedPoints(), this);
+            case RotatingPolygon polygon -> CollisionsUtil.collidesWith(polygon.getRotatedPoints(), this);
             default -> false;
         };
     }
